refactor(hosts): split connection and refhost into subpackages
Step 8 of the reorg plan, part 1 of 2: move the two oversize
top-level modules into the new hosts/ cluster along their
already-present internal seams. target/ moves as a directory in
the following commit.

connection.py (783 LOC) -> mtui/hosts/connection/
  - timeout.py  : CommandTimeoutError + policy_from_config (lines 1-78)
  - connection.py : class Connection (line 80 to end of file)
  - __init__.py re-exports all three public names so callers can
    keep writing 'from mtui.hosts.connection import Connection'.

refhost.py (542 LOC) -> mtui/hosts/refhost/
  - models.py   : Version/Product/Addon/Host/Attributes dataclasses
                  plus the free _format_named_version /
                  _parse_named_version / _attributes_with_arch /
                  _version_from_dict / _host_from_dict helpers.
  - store.py    : class Refhosts (YAML loader + matcher) plus the
                  factory dispatch trio (RefhostsResolveFailedError,
                  _RefhostsFactory). RefhostsResolveFailedError
                  moves to store.py rather than the plan's
                  resolvers.py because it is *raised by* the factory
                  and importing it from resolvers.py would create a
                  store<->resolvers cycle. Documented as a deviation.
  - resolvers.py: Resolver ABC + PathResolver + HttpsResolver.
  - __init__.py builds the RefhostsFactory singleton (the only
    code that needs both halves of the package) and re-exports
    the public surface plus _RefhostsFactory. The plan's letter
    placed the singleton in store.py, but constructing it there
    would need to import resolvers, creating the same cycle.
    Same shape as the data_sources/oqa_search/__init__.py deviation
    (commit fb0a5e3) for whitebox-test surface preservation.

Logger names ('mtui.connection', 'mtui.refhost') are kept at their
original strings per the policy established in commit 1be141b
(support/ move). Six caplog.at_level(logger=...) call sites in
tests/test_connection*.py and tests/test_refhost.py thus keep working
unchanged.

Importers updated this commit:
  - mtui/support/config.py            ..refhost -> ..hosts.refhost
  - mtui/commands/simpleset.py        ..refhost -> ..hosts.refhost
  - mtui/test_reports/testreport.py   ..refhost -> ..hosts.refhost
  - mtui/target/{target,locks}.py     ..connection -> ..hosts.connection
  - mtui/target/parsers/system.py     ...connection -> ...hosts.connection
  - tests/test_refhost.py             'from mtui import refhost'
                                      -> 'from mtui.hosts import refhost'
                                      (six caplog logger= strings unchanged)
  - tests/test_config.py              mtui.refhost -> mtui.hosts.refhost
  - tests/test_testreport.py:343      mtui.refhost -> mtui.hosts.refhost
  - tests/test_target.py:242          mtui.connection -> mtui.hosts.connection
  - tests/test_connection.py          import + 8 monkeypatch/patch strings
                                      ('mtui.connection.<X>' targets the
                                      leaf module attrs: SSHClient/SSHConfig
                                      /Path/select.select, so the new
                                      target is mtui.hosts.connection.connection.<X>)
  - tests/test_connection_deep.py     same shape, 13 sites

Survey: 22 sites total in this commit. The remaining target/ +
target.locks.<X> mock targets land in the next commit when target/
moves into mtui/hosts/.

Gates: ruff format --check, ruff check, ty check all green;
pytest = 989 passed; uv run python -m mtui --help works.
