+++ mtui/commands/simpleset.py
@@ -6,7 +6,7 @@ from ..cli.argparse import ArgumentParser
 from ..cli.completion import complete_choices
 from ..data_sources.openqa import KernelOpenQA
 from ..data_sources.qem_dashboard import DashboardAutoOpenQA
-from ..refhost import RefhostsFactory
+from ..hosts.refhost import RefhostsFactory
 from ..support import messages
 from ..support.misc import requires_update
 from . import Command
