+++ mtui/hosts/connection/connection.py
@@ -22,8 +22,9 @@ from traceback import format_exc
 import paramiko
 from paramiko import Channel, SFTPClient, SFTPFile, SSHClient, SSHConfig
 
-from .cli.term import termsize
-from .support.messages import ReConnectFailed
+from ...cli.term import termsize
+from ...support.messages import ReConnectFailed
+from .timeout import CommandTimeoutError
 
 logger = getLogger("mtui.connection")
 RETRIES: int = 5
@@ -35,46 +36,6 @@ if not sys.warnoptions:
     # Suppress only paramiko/cryptography deprecation warnings, not all warnings
     warnings.filterwarnings("ignore", module="paramiko")
     warnings.filterwarnings("ignore", module="cryptography")
-
-
-class CommandTimeoutError(Exception):
-    """Exception raised when a remote command times out."""
-
-    def __init__(self, command=None) -> None:
-        """Initializes the exception.
-
-        Args:
-            command: The command that timed out.
-
-        """
-        self.command = command
-
-    def __str__(self) -> str:
-        """Returns the timed out remote command as a string."""
-        return repr(self.command)
-
-
-_HOST_KEY_POLICIES: dict[str, type[paramiko.MissingHostKeyPolicy]] = {
-    "auto_add": paramiko.AutoAddPolicy,
-    "warn": paramiko.WarningPolicy,
-    "reject": paramiko.RejectPolicy,
-}
-
-
-def policy_from_config(name: str) -> paramiko.MissingHostKeyPolicy:
-    """Map an ``ssh_strict_host_key_checking`` config value to a paramiko policy.
-
-    Unknown values fall back to ``AutoAddPolicy`` (preserving the legacy
-    behaviour) and emit a warning so misconfigurations are visible.
-    """
-    cls = _HOST_KEY_POLICIES.get(name)
-    if cls is None:
-        logger.warning(
-            "unknown ssh_strict_host_key_checking=%r; falling back to auto_add",
-            name,
-        )
-        cls = paramiko.AutoAddPolicy
-    return cls()
 
 
 class Connection:
