+++ mtui/hosts/connection/timeout.py
@@ -1,0 +1,52 @@
+"""Command-timeout error and host-key policy mapping.
+
+These helpers live alongside :class:`~mtui.hosts.connection.connection.Connection`
+but have no dependency on it, so they sit in their own module to keep
+``connection.py`` focused on the SSH/SFTP wrapper proper.
+"""
+
+from logging import getLogger
+
+import paramiko
+
+logger = getLogger("mtui.connection")
+
+
+class CommandTimeoutError(Exception):
+    """Exception raised when a remote command times out."""
+
+    def __init__(self, command=None) -> None:
+        """Initializes the exception.
+
+        Args:
+            command: The command that timed out.
+
+        """
+        self.command = command
+
+    def __str__(self) -> str:
+        """Returns the timed out remote command as a string."""
+        return repr(self.command)
+
+
+_HOST_KEY_POLICIES: dict[str, type[paramiko.MissingHostKeyPolicy]] = {
+    "auto_add": paramiko.AutoAddPolicy,
+    "warn": paramiko.WarningPolicy,
+    "reject": paramiko.RejectPolicy,
+}
+
+
+def policy_from_config(name: str) -> paramiko.MissingHostKeyPolicy:
+    """Map an ``ssh_strict_host_key_checking`` config value to a paramiko policy.
+
+    Unknown values fall back to ``AutoAddPolicy`` (preserving the legacy
+    behaviour) and emit a warning so misconfigurations are visible.
+    """
+    cls = _HOST_KEY_POLICIES.get(name)
+    if cls is None:
+        logger.warning(
+            "unknown ssh_strict_host_key_checking=%r; falling back to auto_add",
+            name,
+        )
+        cls = paramiko.AutoAddPolicy
+    return cls()
