+++ tests/test_refhost.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from mtui import refhost
+from mtui.hosts import refhost
 from mtui.support.messages import InvalidLocationError
 
 REFHOSTS_FIXTURE = Path(__file__).parent / "fixtures" / "refhosts.yml"
